introduce dangling_cleanup attribute and change source to use it
Some functions leak cleanups on purpose.  However, these functions
aren't "constructors" in the cleanup-checker sense, as they do not
return the cleanup.

This patch adds an attribute that can be used to mark these functions.
This avoids a bunch of false reports from the checker.

This patch is really a huge hack.  It is basically giving up on
cleanly checking the use of cleanups.  You can see this most easily by
noticing that I elected not to have this property propagate through
the call chain.

Still, I figured I'd see what people thought of this patch.  Without
it the checker is harder to use, as you must know which functions are
ignorable.

Maybe some of these are fixable with a bit more thought.

	* breakpoint.c (breakpoint_re_set_default): Use DANGLING_CLEANUP.
	* c-exp.y (operator_stoken): Use DANGLING_CLEANUP.
	* coffread.c (coff_locate_sections): Use DANGLING_CLEANUP.
	* contrib/cleanup_check.py (leaks_cleanup): New global.
	(note_dangling_cleanup, leaves_dangling_cleanup): New functions.
	(CleanupChecker.compute_master): Add comment.  Check
	leaves_dangling_cleanup.
	(CleanupChecker.check_cleanups): Don't mention special constructors
	to the user.
	(register_attributes): New function.  Register it with gcc.
	* defs.h (DANGLING_CLEANUP): New define.
	* dwarf2read.c (read_cutu_die_from_dwo, handle_DW_AT_stmt_list)
	(setup_type_unit_groups, dwarf2_add_field, dwarf2_add_typedef)
	(dwarf2_add_member_fn, psymtab_include_file_name)
	(dwarf_decode_lines): Use DANGLING_CLEANUP.
	* linux-nat.c (linux_child_pid_to_exec_file): Use DANGLING_CLEANUP.
	* linux-tdep.c (linux_make_corefile_notes): Use DANGLING_CLEANUP.
	* python/py-prettyprint.c (push_dummy_python_frame): Use
	DANGLING_CLEANUP.
	* stabsread.c (read_member_functions, read_struct_fields)
	(read_baseclasses): Use DANGLING_CLEANUP.
	* utils.h (gdb_bfd_errmsg): Use DANGLING_CLEANUP.
	* tracepoint.c (trace_dump_actions): Use DANGLING_CLEANUP.