use ptid_build_target in more places
This adds uses of ptid_build_target in more places.
This fixes a few regressions.

It also changes how target stack IDs are handed out,
to ensure that unconverted targets always get a stack ID
of 0.  This lets ptid_build continue to work ok.

This bit seems like a hack, we should probably just do a full
conversion.

Also see the hacky defines of target_stack_id in common/.

One idea would be to move this sort of hack to ptid.h
or even bury it in ptid.c, which is arguably where it belongs.