remove simple GDBSERVER uses from common
This removes various simple GDBSERVER uses from common.  The simple
uses are just cases where the code includes defs.h or server.h
depending on GDBSERVER.  Instead, now the files include the headers
that they require.  Unfortunately for the time being we still need to
check GDBSERVER to decide which gnulib config header to import; but
this is not very significant in the long run.

2014-06-18  Tom Tromey  <tromey@redhat.com>

	* common/buffer.c: Don't include server.h or defs.h; update
	includes.
	* common/common-utils.c: Don't include server.h or defs.h; update
	includes.
	* common/filestuff.c: Don't include server.h or defs.h; update
	includes.
	* common/filestuff.h: Include stdio.h.
	* common/format.c: Don't include server.h or defs.h; update
	includes.
	* common/gdb_vecs.c: Don't include server.h or defs.h; update
	includes.
	* common/linux-osdata.c: Don't include server.h or defs.h; update
	includes.
	* common/linux-procfs.c: Don't include server.h or defs.h; update
	includes.
	* common/signals.c: Don't include server.h or defs.h; update
	includes.
	* common/vec.c: Don't include server.h or defs.h; update includes.
	* common/xml-utils.c: Don't include server.h or defs.h; update
	includes.