remove to_data
There's no reason to have to_data.

On the one hand, a to_xclose target can readily subclass target_ops.
This is simple and clean.

On the other hand, a to_close target can't really use to_data in the
multi-target future, because such a target is inherently not
multi-capable.

So, this patch removes the field and fixes up its sole user.

2014-07-29  Tom Tromey  <tromey@redhat.com>

	* bfd-target.c (struct target_bfd_ops): Rename from
	target_bfd_data.  Add "base" field.
	(target_bfd_xfer_partial, target_bfd_get_section_table)
	(target_bfd_xclose, target_bfd_reopen): Update.
	* target.h (struct target_ops) <to_data>: Remove.