Enable libstdc++ debug mode in gdb
Defining _GLIBCXX_DEBUG caught a bug earlier, and I think it makes
sense to enable it by default when in development mode, in the hopes
of preventing other bugs.

Because the libstdc++ development mode can come with a performance
cost, I've added the ability to disable it with a configure option.
Additionally, it is disabled by default in release builds.

gdb/ChangeLog
2018-07-31  Tom Tromey  <tom@tromey.com>

	* cxx-debug.m4: New file.
	* configure.ac: Use AM_GDB_CXX_DEBUG.
	* config.in, configure: Rebuild.
	* acinclude.m4: Include cxx-debug.m4.

gdb/gdbserver/ChangeLog
2018-07-31  Tom Tromey  <tom@tromey.com>

	* configure.ac: Use AM_GDB_CXX_DEBUG.
	* config.in, configure: Rebuild.
	* acinclude.m4: Include ../cxx-debug.m4.